refactor: #5 기능별 파일 구조 정리
