+++ app/src/main/java/com/sopt/now/data/Friend.kt
@@ -1,4 +1,4 @@
-package com.sopt.now
+package com.sopt.now.data
 
 import androidx.annotation.DrawableRes
