+++ app/src/main/java/com/sopt/now/data/User.kt
@@ -1,4 +1,4 @@
-package com.sopt.now
+package com.sopt.now.data
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
