+++ app/src/main/java/com/sopt/now/ui/home/UserViewHolder.kt
@@ -1,6 +1,7 @@
-package com.sopt.now
+package com.sopt.now.ui.home
 
 import androidx.recyclerview.widget.RecyclerView
+import com.sopt.now.data.User
 import com.sopt.now.databinding.ItemUserBinding
 
 class UserViewHolder(private val binding: ItemUserBinding) : RecyclerView.ViewHolder(binding.root) {
