+++ app/src/main/java/com/sopt/now/ui/main/MainActivity.kt
@@ -1,9 +1,14 @@
-package com.sopt.now
+package com.sopt.now.ui.main
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
+import com.sopt.now.R
+import com.sopt.now.data.User
 import com.sopt.now.databinding.ActivityMainBinding
+import com.sopt.now.ui.home.HomeFragment
+import com.sopt.now.ui.mypage.MyPageFragment
+import com.sopt.now.ui.search.SearchFragment
 
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
