+++ app/src/main/java/com/sopt/now/ui/search/SearchFragment.kt
@@ -1,10 +1,11 @@
-package com.sopt.now
+package com.sopt.now.ui.search
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.sopt.now.R
 
 class SearchFragment : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
