+++ app/src/main/res/layout/activity_sign_up.xml
@@ -5,7 +5,7 @@
     android:id="@+id/layout_signup"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SignUpActivity">
+    tools:context=".ui.signup.SignUpActivity">
 
     <TextView
         android:id="@+id/tv_signup_title"
